+++ docs/releasenotes.rst
@@ -12,6 +12,7 @@ V2.0.9:
 
 - fix: Python parser - header detection correction for nmon external monitoring
 - feature: Add df information for improved file system monitoring and storage capacity planning
+- fix: Perl parser issue - UARG parsing issue for AIX #5
 
 =======
 V2.0.8: